feat: Tandem becomes it's own project.
Tandem is a command line tool for updating the Discord Linux desktop
application. The main bulk of the work is done by wrapping dpkg.

Tandem started off as a standalone Python script within the provisioning
project, and quickly expanded in scope.

Given that the Provisioning project is more about configuration
management than actual tool implementations, it made sense to break this
into it's own project, and then Provisioning will pull in a release
binary.

Documentation:
* https://github.com/Darianisak/provisioning